Break up test_utils.rs: deploy account tx
commit-id:67f489ab
